pyproject.toml
[project]
name = "slack-ai-bot"
version = "0.1.0"
description = "Slack bot that integrates with multiple AI services"
dependencies = [
    "slack-bolt>=1.18.1",
    "openai>=1.12.0",
    "anthropic>=0.18.1",
    "google-generativeai>=0.3.2",
    "google-cloud-firestore>=2.15.0",
    "python-dotenv>=1.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
